+++ spec/bind-query.js
@@ -1,6 +1,6 @@
 var assert = require('chai').assert,
-  bindQuery = require('../../bin/bind-query'),
-  config = require('../../config.json'),
+  bindQuery = require('../bin/bind-query'),
+  config = require('../config.json'),
   Spinner = require('cli-spinner').Spinner,
   spinner = new Spinner('processing... %s');
